+++ src/__tests__/config-loader.test.ts
@@ -1,10 +1,9 @@
-import { assert } from "chai";
 import {
   configLoader,
   loadConfig,
   ConfigLoaderFailResult,
   ConfigLoaderSuccessResult,
-} from "../src/config-loader";
+} from "../config-loader";
 import { join } from "path";
 
 describe("config-loader", (): void => {
@@ -20,9 +19,9 @@ describe("config-loader", (): void => {
     });
 
     const successResult = result as ConfigLoaderSuccessResult;
-    assert.equal(successResult.resultType, "success");
-    assert.equal(successResult.absoluteBaseUrl, "/foo/bar");
-    assert.equal(successResult.paths["asd"][0], "asd");
+    expect(successResult.resultType).toBe("success");
+    expect(successResult.absoluteBaseUrl).toBe("/foo/bar");
+    expect(successResult.paths["asd"][0]).toBe("asd");
   });
 
   it("should use explicitParams when set and add cwd when path is relative", () => {
@@ -37,16 +36,15 @@ describe("config-loader", (): void => {
     });
 
     const successResult = result as ConfigLoaderSuccessResult;
-    assert.equal(successResult.resultType, "success");
-    assert.equal(successResult.absoluteBaseUrl, join("/baz", "bar/"));
+    expect(successResult.resultType).toBe("success");
+    expect(successResult.absoluteBaseUrl).toBe(join("/baz", "bar/"));
   });
 
   it("should fallback to tsConfigLoader when explicitParams is not set", () => {
     const result = configLoader({
       explicitParams: undefined,
       cwd: "/baz",
-      // tslint:disable-next-line:no-any
-      tsConfigLoader: (_: any) => ({
+      tsConfigLoader: () => ({
         tsConfigPath: "/baz/tsconfig.json",
         baseUrl: "./src",
         paths: {},
@@ -54,16 +52,15 @@ describe("config-loader", (): void => {
     });
 
     const successResult = result as ConfigLoaderSuccessResult;
-    assert.equal(successResult.resultType, "success");
-    assert.equal(successResult.absoluteBaseUrl, join("/baz", "src"));
+    expect(successResult.resultType).toBe("success");
+    expect(successResult.absoluteBaseUrl).toBe(join("/baz", "src"));
   });
 
-  it("should show an error message when baseUrl is missing", () => {
+  it("should tolerate a missing baseUrl", () => {
     const result = configLoader({
       explicitParams: undefined,
       cwd: "/baz",
-      // tslint:disable-next-line:no-any
-      tsConfigLoader: (_: any) => ({
+      tsConfigLoader: () => ({
         tsConfigPath: "/baz/tsconfig.json",
         baseUrl: undefined,
         paths: {},
@@ -71,8 +68,7 @@ describe("config-loader", (): void => {
     });
 
     const failResult = result as ConfigLoaderFailResult;
-    assert.equal(failResult.resultType, "failed");
-    assert.isTrue(failResult.message.indexOf("baseUrl") > -1);
+    expect(failResult.resultType).toBe("success");
   });
 
   it("should presume cwd to be a tsconfig file when loadConfig is called with absolute path to tsconfig.json", () => {
@@ -84,7 +80,22 @@ describe("config-loader", (): void => {
     const result = loadConfig(configFile);
 
     const successResult = result as ConfigLoaderSuccessResult;
-    assert.equal(successResult.resultType, "success");
-    assert.equal(successResult.configFileAbsolutePath, configFile);
+    expect(successResult.resultType).toBe("success");
+    expect(successResult.configFileAbsolutePath).toBe(configFile);
+  });
+
+  it("should allow an absolute baseUrl in tsconfig.json", () => {
+    const result = configLoader({
+      explicitParams: undefined,
+      cwd: "/baz",
+      tsConfigLoader: () => ({
+        tsConfigPath: "/baz/tsconfig.json",
+        baseUrl: "/baz",
+        paths: {},
+      }),
+    });
+
+    const successResult = result as ConfigLoaderSuccessResult;
+    expect(successResult.absoluteBaseUrl).toEqual("/baz");
   });
 });