+++ src/__tests__/match-path-async.test.ts
@@ -1,5 +1,4 @@
-import { assert } from "chai";
-import { createMatchPathAsync } from "../src/match-path-async";
+import { createMatchPathAsync } from "../match-path-async";
 import * as Tests from "./data/match-path-data";
 
 describe("match-path-async", () => {
@@ -18,7 +17,7 @@ describe("match-path-async", () => {
           callback(undefined, t.existingFiles.indexOf(path) !== -1),
         t.extensions,
         (_err, result) => {
-          assert.equal(result, t.expectedPath);
+          expect(result).toBe(t.expectedPath);
           done();
         }
       );