+++ src/__tests__/try-path.test.ts
@@ -1,5 +1,4 @@
-import { assert } from "chai";
-import { getPathsToTry } from "../src/try-path";
+import { getPathsToTry } from "../try-path";
 import { join } from "path";
 
 describe("mapping-entry", () => {
@@ -11,13 +10,23 @@ describe("mapping-entry", () => {
     { pattern: "pre/fix/*", paths: [join("/absolute", "base", "url", "foo3")] },
     { pattern: "*", paths: [join("/absolute", "base", "url", "foo1")] },
   ];
+  const abosolutePathMappingsStarstWithSlash = [
+    {
+      pattern: "/opt/*",
+      paths: [join("/absolute", "src", "aws-layer")],
+    },
+    {
+      pattern: "*",
+      paths: [join("/absolute", "src")],
+    },
+  ];
   it("should return no paths for relative requested module", () => {
     const result = getPathsToTry(
       [".ts", "tsx"],
       abosolutePathMappings,
       "./requested-module"
     );
-    assert.deepEqual(result, undefined);
+    expect(result).toBeUndefined();
   });
 
   it("should return no paths if no pattern match the requested module", () => {
@@ -35,7 +44,7 @@ describe("mapping-entry", () => {
       ],
       "requested-module"
     );
-    assert.deepEqual(result, undefined);
+    expect(result).toBeUndefined();
   });
 
   it("should get all paths that matches requested module", () => {
@@ -44,7 +53,7 @@ describe("mapping-entry", () => {
       abosolutePathMappings,
       "longest/pre/fix/requested-module"
     );
-    assert.deepEqual(result, [
+    expect(result).toEqual([
       // "longest/pre/fix/*"
       { type: "file", path: join("/absolute", "base", "url", "foo2", "bar") },
       {
@@ -85,15 +94,48 @@ describe("mapping-entry", () => {
       },
     ]);
   });
-});
 
-// describe("match-star", () => {
-//   it("should match star in last position", () => {
-//     const result = matchStar("lib/*", "lib/mylib");
-//     assert.equal(result, "mylib");
-//   });
-//   it("should match star in first position", () => {
-//     const result = matchStar("*/lib", "mylib/lib");
-//     assert.equal(result, "mylib");
-//   });
-// });
+  it("should resolve paths starting with a slash", () => {
+    const result = getPathsToTry(
+      [".ts"],
+      abosolutePathMappingsStarstWithSlash,
+      "/opt/utils"
+    );
+    expect(result).toEqual([
+      // "opt/*"
+      {
+        path: join("/absolute", "src", "aws-layer"),
+        type: "file",
+      },
+      {
+        path: join("/absolute", "src", "aws-layer.ts"),
+        type: "extension",
+      },
+      {
+        path: join("/absolute", "src", "aws-layer", "package.json"),
+        type: "package",
+      },
+      {
+        path: join("/absolute", "src", "aws-layer", "index.ts"),
+        type: "index",
+      },
+      // "*"
+      {
+        path: join("/absolute", "src"),
+        type: "file",
+      },
+      {
+        path: join("/absolute", "src.ts"),
+        type: "extension",
+      },
+      {
+        path: join("/absolute", "src", "package.json"),
+        type: "package",
+      },
+      {
+        path: join("/absolute", "src", "index.ts"),
+        type: "index",
+      },
+    ]);
+  });
+});