+++ src/__tests__/tsconfig-loader.test.ts
@@ -1,9 +1,8 @@
-import { assert } from "chai";
 import {
   loadTsconfig,
   tsConfigLoader,
   walkForTsConfig,
-} from "../src/tsconfig-loader";
+} from "../tsconfig-loader";
 import { join } from "path";
 
 describe("tsconfig-loader", () => {
@@ -20,7 +19,7 @@ describe("tsconfig-loader", () => {
       },
     });
 
-    assert.equal(result.tsConfigPath, "/foo/bar/tsconfig.json");
+    expect(result.tsConfigPath).toBe("/foo/bar/tsconfig.json");
   });
 
   it("should return loaderResult.tsConfigPath as undefined when not found", () => {
@@ -36,7 +35,7 @@ describe("tsconfig-loader", () => {
       },
     });
 
-    assert.isUndefined(result.tsConfigPath);
+    expect(result.tsConfigPath).toBeUndefined();
   });
 
   it("should use TS_NODE_PROJECT env if exists", () => {
@@ -61,32 +60,110 @@ describe("tsconfig-loader", () => {
       },
     });
 
-    assert.equal(result.tsConfigPath, "/foo/baz/tsconfig.json");
+    expect(result.tsConfigPath).toBe("/foo/baz/tsconfig.json");
+  });
+
+  it("should use TS_NODE_BASEURL env if exists", () => {
+    const result = tsConfigLoader({
+      cwd: "/foo/bar",
+      getEnv: (key: string) =>
+        key === "TS_NODE_BASEURL" ? "SOME_BASEURL" : undefined,
+      loadSync: (_0: string, _1: string, baseUrl: string) => {
+        return {
+          tsConfigPath: undefined,
+          baseUrl,
+          paths: {},
+        };
+      },
+    });
+
+    expect(result.baseUrl).toBe("SOME_BASEURL");
+  });
+
+  it("should not use TS_NODE_BASEURL env if it does not exist", () => {
+    const result = tsConfigLoader({
+      cwd: "/foo/bar",
+      getEnv: (_: string) => {
+        return undefined;
+      },
+      loadSync: (_0: string, _1: string, baseUrl: string) => {
+        return {
+          tsConfigPath: undefined,
+          baseUrl,
+          paths: {},
+        };
+      },
+    });
+
+    expect(result.baseUrl).toBeUndefined();
   });
 });
 
 describe("walkForTsConfig", () => {
   it("should find tsconfig in starting directory", () => {
     const pathToTsconfig = join("/root", "dir1", "tsconfig.json");
+    const mockFiles: Record<string, string[]> = {
+      "/root/dir1": ["tsconfig.json"],
+    };
     const res = walkForTsConfig(
       join("/root", "dir1"),
-      (path) => path === pathToTsconfig
+      (path) => mockFiles[path] || []
     );
-    assert.equal(res, pathToTsconfig);
+    expect(res).toBe(pathToTsconfig);
+  });
+
+  it("should find jsconfig in starting directory", () => {
+    const pathToJsconfig = join("/root", "dir1", "jsconfig.json");
+    const mockFiles: Record<string, string[]> = {
+      "/root/dir1": ["jsconfig.json"],
+    };
+    const res = walkForTsConfig(
+      join("/root", "dir1"),
+      (path) => mockFiles[path] || []
+    );
+    expect(res).toBe(pathToJsconfig);
+  });
+
+  // see https://github.com/Microsoft/TypeScript/issues/15869#issuecomment-301845650
+  it("tsconfig.json take precedence over jsconfig.json when both exist", () => {
+    const pathToTsconfig = join("/root/dir1", "tsconfig.json");
+    const mockFiles: Record<string, string[]> = {
+      "/root/dir1": ["jsconfig.json", "tsconfig.json"],
+    };
+    const res = walkForTsConfig(
+      join("/root", "dir1"),
+      (path) => mockFiles[path] || []
+    );
+    expect(res).toBe(pathToTsconfig);
   });
 
   it("should find tsconfig in parent directory", () => {
     const pathToTsconfig = join("/root", "tsconfig.json");
+    const mockFiles: Record<string, string[]> = {
+      "/root": ["tsconfig.json"],
+    };
     const res = walkForTsConfig(
       join("/root", "dir1"),
-      (path) => path === pathToTsconfig
+      (path) => mockFiles[path] || []
     );
-    assert.equal(res, pathToTsconfig);
+    expect(res).toBe(pathToTsconfig);
+  });
+
+  it("should find jsconfig in parent directory", () => {
+    const pathToTsconfig = join("/root", "jsconfig.json");
+    const mockFiles: Record<string, string[]> = {
+      "/root": ["jsconfig.json"],
+    };
+    const res = walkForTsConfig(
+      join("/root", "dir1"),
+      (path) => mockFiles[path] || []
+    );
+    expect(res).toBe(pathToTsconfig);
   });
 
   it("should return undefined when reaching the top", () => {
-    const res = walkForTsConfig(join("/root", "dir1", "kalle"), () => false);
-    assert.equal(res, undefined);
+    const res = walkForTsConfig(join("/root", "dir1", "kalle"), () => []);
+    expect(res).toBeUndefined();
   });
 });
 
@@ -98,7 +175,7 @@ describe("loadConfig", () => {
       (path) => path === "/root/dir1/tsconfig.json",
       (_) => JSON.stringify(config)
     );
-    assert.deepEqual(res, config);
+    expect(res).toStrictEqual(config);
   });
 
   it("It should load a config with comments", () => {
@@ -113,7 +190,7 @@ describe("loadConfig", () => {
           }
         }`
     );
-    assert.deepEqual(res, config);
+    expect(res).toStrictEqual(config);
   });
 
   it("It should load a config with trailing commas", () => {
@@ -127,7 +204,7 @@ describe("loadConfig", () => {
           },
         }`
     );
-    assert.deepEqual(res, config);
+    expect(res).toStrictEqual(config);
   });
 
   it("It should throw an error including the file path when encountering invalid JSON5", () => {
@@ -172,7 +249,7 @@ describe("loadConfig", () => {
       }
     );
 
-    assert.deepEqual(res, {
+    expect(res).toEqual({
       extends: "../base-config.json",
       compilerOptions: {
         baseUrl: "kalle",
@@ -216,7 +293,7 @@ describe("loadConfig", () => {
       }
     );
 
-    assert.deepEqual(res, {
+    expect(res).toEqual({
       extends: "my-package/base-config.json",
       compilerOptions: {
         baseUrl: "kalle",
@@ -253,7 +330,7 @@ describe("loadConfig", () => {
       }
     );
 
-    assert.deepEqual(res, {
+    expect(res).toEqual({
       extends: "../second-config.json",
       compilerOptions: { baseUrl: join("..", "..") },
     });