+++ src/match-path-async.ts
@@ -17,6 +17,7 @@ export interface MatchPathAsync {
 }
 
 export interface MatchPathAsyncCallback {
+  // eslint-disable-next-line no-shadow
   (err?: Error, path?: string): void;
 }
 
@@ -97,7 +98,7 @@ function findFirstExistingMainFieldMappedFile(
     return doneCallback(undefined, undefined);
   }
 
-  const tryNext = () =>
+  const tryNext = (): void =>
     findFirstExistingMainFieldMappedFile(
       packageJson,
       mainFields,
@@ -148,7 +149,6 @@ function findFirstExistingPath(
         return doneCallback(err);
       }
       if (exists) {
-        // Not sure why we don't just return the full path? Why strip it?
         return doneCallback(undefined, TryPath.getStrippedPath(tryPath));
       }
       if (index === tryPaths.length - 1) {
@@ -180,11 +180,7 @@ function findFirstExistingPath(
               return doneCallback(mainFieldErr);
             }
             if (mainFieldMappedFile) {
-              // Not sure why we don't just return the full path? Why strip it?
-              return doneCallback(
-                undefined,
-                Filesystem.removeExtension(mainFieldMappedFile)
-              );
+              return doneCallback(undefined, mainFieldMappedFile);
             }
 
             // No field in package json was a valid option. Continue with the next path.