+++ src/register.ts
@@ -1,6 +1,5 @@
 import { createMatchPath } from "./match-path-sync";
 import { configLoader, ExplicitParams } from "./config-loader";
-import { options } from "./options";
 
 const noOp = (): void => void 0;
 
@@ -45,13 +44,40 @@ function getCoreModules(
   return coreModules;
 }
 
+export interface RegisterParams extends ExplicitParams {
+  /**
+   * Defaults to `--project` CLI flag or `process.cwd()`
+   */
+  cwd?: string;
+}
+
 /**
  * Installs a custom module load function that can adhere to paths in tsconfig.
  * Returns a function to undo paths registration.
  */
-export function register(explicitParams: ExplicitParams): () => void {
+export function register(params?: RegisterParams): () => void {
+  let cwd: string | undefined;
+  let explicitParams: ExplicitParams | undefined;
+  if (params) {
+    cwd = params.cwd;
+    if (params.baseUrl || params.paths) {
+      explicitParams = params;
+    }
+  } else {
+    // eslint-disable-next-line
+    const minimist = require("minimist");
+    const argv = minimist(process.argv.slice(2), {
+      // eslint-disable-next-line id-denylist
+      string: ["project"],
+      alias: {
+        project: ["P"],
+      },
+    });
+    cwd = argv.project;
+  }
+
   const configLoaderResult = configLoader({
-    cwd: options.cwd,
+    cwd: cwd ?? process.cwd(),
     explicitParams,
   });
 
@@ -71,27 +97,27 @@ export function register(explicitParams: ExplicitParams): () => void {
   );
 
   // Patch node's module loading
-  // tslint:disable-next-line:no-require-imports variable-name
+  // eslint-disable-next-line @typescript-eslint/no-require-imports,@typescript-eslint/no-var-requires
   const Module = require("module");
+  // eslint-disable-next-line no-underscore-dangle
   const originalResolveFilename = Module._resolveFilename;
   const coreModules = getCoreModules(Module.builtinModules);
-  // tslint:disable-next-line:no-any
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any,no-underscore-dangle
   Module._resolveFilename = function (request: string, _parent: any): string {
     const isCoreModule = coreModules.hasOwnProperty(request);
     if (!isCoreModule) {
       const found = matchPath(request);
       if (found) {
         const modifiedArguments = [found, ...[].slice.call(arguments, 1)]; // Passes all arguments. Even those that is not specified above.
-        // tslint:disable-next-line:no-invalid-this
         return originalResolveFilename.apply(this, modifiedArguments);
       }
     }
-    // tslint:disable-next-line:no-invalid-this
     return originalResolveFilename.apply(this, arguments);
   };
 
   return () => {
     // Return node's module loading to original state.
+    // eslint-disable-next-line no-underscore-dangle
     Module._resolveFilename = originalResolveFilename;
   };
 }